Move Tag component into molecules
