+++ src/components/TrackItem.tsx
@@ -4,7 +4,7 @@ import { css } from "g-style";
 import Breakpoints from "components/atoms/Breakpoints";
 
 import Link from "components/Link";
-import Tag from "components/Tag";
+import Tag from "components/molecules/Tag";
 
 const className = css({
   flexDirection: "column",
